Crate e2e_tests

Test infrastructure to stand up a Sui localnet, a bitcoin regtest, and hashi nodes.

The general bootstrapping process is as follows:

Stand up a Bitcoin regtest

Stand up a Sui Network leveraging sui start.

Ensure that the SuiSystemState object has been upgraded from v1 to v2.

Ensure that each sui validator address is properly funded.

Publish the Hashi package.

Build configs for each Hashi node (one for each validator).

Register each validator with the Hashi system object

Initialize the first hashi committee once all validators have been registered.

Re-exports

pub use bitcoin_node::BitcoinNodeBuilder;

pub use bitcoin_node::BitcoinNodeHandle;

pub use hashi_network::HashiNetwork;

pub use hashi_network::HashiNetworkBuilder;

pub use hashi_network::HashiNodeHandle;

pub use sui_network::SuiNetworkBuilder;

pub use sui_network::SuiNetworkHandle;

Modules

backup_restore

End-to-end backup/restore round-trip test.

bitcoin_node

e2e_flow

hashi_network

sui_network

test_helpers

Shared helpers used by e2e test modules.

upgrade_flow

Upgrade test infrastructure.

Structs

TestNetworks

TestNetworksBuilder