e2e_tests::steps::jsonrpc_call
Function process_json_variables

fn process_json_variables(
    value: &mut Value,
    ctx: &TestContext<'_>,
) -> Result<()>

Recursively processes a JSON value, replacing variable references like ${variable_name} with corresponding values from the TestContext.

Used internally by CallStep to:

Substitute variables in JSON RPC input arguments before making the call

Process expected result templates to compare with actual responses

For example:

Input args: {“proposalId”: “${proposal_id}”}

Expected result: {“status”: “${proposal_id}”}

Arguments

value - JSON value to process, modified in place

ctx - Test context containing variable values

Errors

If a referenced variable is not found in the context